2단계 - 추가 기능 및 리팩터링 (#84)
* feat first commit

* feat 자동차 경주 뼈대 코드 준비

* refactor 기존 코드를 Spring에 맞도록 수정 및 Controller에 맞는 코드로 변경

* refactor 기존 코드에서 Spring에서 사용하지 않는 클래스들 삭제

* feat Spring에서 racing car에 대한 정보를 request하는 class 추가

* feat Spring에서 racing car에 대한 정보를 response 받는 class 추가

* refactor resource에 맞는 carName변수명을 name으로 변경

* feat 기존 RacingCar.class의 코드를 RacingCarService로 이동 및 DB에서 받아오는 값에 맞게 변경

* style 줄바꿈 정리

* feat GameHistoryDao 추가

* feat GameResultDao 추가

* feat GameHistory 추가

* feat GameResult 추가

* feat GameHistoryRepository 추가

* feat GameResultRepository 추가

* feat RacingCarService에서 game관련된 함수 분리

* refactor 게임 관련 기능 이동으로 인해 코드 수정

* feat data resource 추가

* style 줄바꿈 정리

* feat gradle 추가

* refactor 안쓰는 클래스 및 메소드 삭제

* refactor 메소드명 수정

* refactor static 삭제

* feat controller에 gamehistory 저장 및 기본 틀 ResponseEntity로 변경

* feat JdbcGameHistoryDao 추가

* feat JdbcGameResultDao 추가

* feat findAll 추가

* feat MemoryGameHistoryDao 추가

* feat MemoryGameResultDao 추가

* refactor 안쓰는 클래스 및 메소드 삭제

* refactor id int에서 long으로 변경

* refactor car클래스의 position초기화 위치 이동

* refactor strategy위치 차 생성시 만들어지도록 위치 변경

* refactor MovingStrategy위치 이동으로 인한 주변 메소드 인자값 변경

* feat RacingCarResponseDto로 변환시키는 메소드 추가

* feat history에 정보를 저장하도록 게임 메소드 변경

* feat GameHistoryResponseDto 추가

* refactor 라이브러리 추가 및 기타 변경

* feat history 정보 load 메소드 추가

* feat 위치 이동 및 사용하지 않는 클래스 삭제

* feat 기존 뼈대 추가 및 Spring 버전에 맞게 변형

* feat 필요없는 annotation 삭제

* refactor join으로 각각의 gameId 에 맞는 정보들을 매칭시켜 조회하도록 변경

* refactor repository package로 이동

* feat 삭제했던 MemoryGameHistoryDao,MemoryGameResultDao class추가

* feat findAll대신 findAllWithGameResults로 변경

* feat 내용 변경

* refactor for문 단순화

* style 줄정리 및 사용하지 않는 메소드 삭제

* feat 추가한 repository관련 메소드 추가

* bugfix jdbc에 null값으로 변경

* bugfix 빈 생성자 삭제

* bugfix 우선순위 지정

* style 줄정리 및 코드 정리

* feat delete test

* feat fix pre-branch

---------

+++ src/main/java/racingcar/domain/GameHistory.java
@@ -3,14 +3,14 @@ package racingcar.domain;
 import java.time.LocalDateTime;
 
 public class GameHistory {
-	private final int id;
-	private final int playResultId;
+	private final long id;
+	private final long playResultId;
 	private final String name;
 	private final int position;
 
 	private final LocalDateTime createdAt;
 
-	public GameHistory(int id, int playResultId, String name, int position, LocalDateTime createdAt) {
+	public GameHistory(long id, long playResultId, String name, int position, LocalDateTime createdAt) {
 		this.id = id;
 		this.playResultId = playResultId;
 		this.name = name;
@@ -18,7 +18,7 @@ public class GameHistory {
 		this.createdAt = createdAt;
 	}
 
-	public static GameHistory of(int gameResultId, Car car) {
+	public static GameHistory of(long gameResultId, Car car) {
 		return GameHistory.builder()
 			.playResultId(gameResultId)
 			.name(car.getName())
@@ -30,11 +30,11 @@ public class GameHistory {
 		return new GameHistoryBuilder();
 	}
 
-	public int getId() {
+	public long getId() {
 		return id;
 	}
 
-	public int getPlayResultId() {
+	public long getPlayResultId() {
 		return playResultId;
 	}
 
@@ -51,19 +51,19 @@ public class GameHistory {
 	}
 
 	public static class GameHistoryBuilder {
-		private int id;
-		private int playResultId;
+		private long id;
+		private long playResultId;
 		private String name;
 		private int position;
 
 		private LocalDateTime createdAt;
 
-		public GameHistoryBuilder id(int id) {
+		public GameHistoryBuilder id(long id) {
 			this.id = id;
 			return this;
 		}
 
-		public GameHistoryBuilder playResultId(int playResultId) {
+		public GameHistoryBuilder playResultId(long playResultId) {
 			this.playResultId = playResultId;
 			return this;
 		}
